Improving error message in DataIO
closes #6673
